scraper_Scripts/requirements.txt
APScheduler==3.9.1
Flask==2.2.2
mysql-connector-python==8.0.30
python-dotenv==0.21.0
requests==2.28.1
requests_html==0.10.0
flask-cors==3.0.10
